+++ proposals/description/fip_104.ts
@@ -1,7 +1,7 @@
 import { ProposalDescription } from '@custom-types/types';
 
-const fip_105: ProposalDescription = {
-  title: 'FIP-105: Reinforce PCV by consolidating assets and perform technical maintenance',
+const fip_104: ProposalDescription = {
+  title: 'FIP-104: Reinforce PCV by consolidating assets and perform technical maintenance',
   commands: [
     {
       target: 'pcvGuardianNew',
@@ -121,7 +121,7 @@ const fip_105: ProposalDescription = {
     }
   ],
   description: `
-  FIP-105: Reinforce PCV by consolidating assets and perform technical maintenance.
+  FIP-104: Reinforce PCV by consolidating assets and perform technical maintenance.
 
   This FIP implements parts of the PCV reinforcement proposal that was approved in this snapshot:
   https://snapshot.fei.money/#/proposal/0x2fd5bdda0067098f6c0520fe309dfe90ca403758f0ce98c1854a00bf38999674 
@@ -144,4 +144,4 @@ const fip_105: ProposalDescription = {
   `
 };
 
-export default fip_105;
+export default fip_104;
