+++ test/integration/proposals_config.ts
@@ -1,13 +1,13 @@
 import { ProposalCategory, ProposalsConfigMap } from '@custom-types/types';
 
-import fip_105 from '@proposals/description/fip_105';
+import fip_104 from '@proposals/description/fip_104';
 // import fip_xx_proposal from '@proposals/description/fip_xx';
 
 const proposals: ProposalsConfigMap = {
-  fip_105: {
+  fip_104: {
     deploy: false, // deploy flag for whether to run deploy action during e2e tests or use mainnet state
     totalValue: 0, // amount of ETH to send to DAO execution
-    proposal: fip_105, // full proposal file, imported from '@proposals/description/fip_xx.ts'
+    proposal: fip_104, // full proposal file, imported from '@proposals/description/fip_xx.ts'
     proposalId: '',
     affectedContractSignoff: [
       'daiFixedPricePSMFeiSkimmer',
